// server/src/modules/access-control/enums/permissions.enum.ts
export enum PermissionsEnum {
  UserGetAll = 'user:get-all:any',
  UserGetAny = 'user:get:any',
  UserCreateAny = 'user:create:any',
  UserUpdateAny = 'user:update:any',
  UserDeleteAny = 'user:delete:any',

  BotGetAll = 'bot:get-all:any',
  BotGetAllByUserAny = 'bot:get-all-by-user:any',
  BotGetAllByUserOwn = 'bot:get-all-by-user:own',
  BotGetAny = 'bot:get:any',
  BotGetOwn = 'bot:get:own',
  BotCreateAny = 'bot:create:any',
  BotCreateOwn = 'bot:create:own',
  BotUpdateAny = 'bot:update:any',
  BotUpdateOwn = 'bot:update:own',
  BotDeleteAny = 'bot:delete:any',
  BotDeleteOwn = 'bot:delete:onw',

  InventoryGetAllByUserAny = 'inventory:get-all-by-user:any',
  InventoryGetAllByUserOwn = 'inventory:get-all-by-user:own',
  InventoryGetAllByBotAny = 'inventory:get-all-by-bot:any',
  InventoryGetAllByBotOwn = 'inventory:get-all-by-bot:own',
  InventoryGetAny = 'inventory:get:any',
  InventoryGetOwn = 'inventory:get:own',
  InventoryCreateAny = 'inventory:create:any',
  InventoryCreateOwn = 'inventory:create:own',
  InventoryRefreshAny = 'inventory:refresh:any',
  InventoryRefreshOwn = 'inventory:refresh:own',
  InventoryDeleteAny = 'inventory:delete:any',
  InventoryDeleteOwn = 'inventory:delete:own',

  ItemGetAll = 'item:get-all:any',
  ItemGetAny = 'item:get:any',
  ItemCreateAny = 'item:create:any',
  ItemUpdateAny = 'item:update:any',
  ItemDeleteAny = 'item:delete:any',
}